irqchip: qcom: mpm: Wire up irq_set_wake call, remove SKIP_SET_WAKE
On Qualcomm SoCs featuring at GICv3 (and maybe also QGIC2), we
need to at least pass through the irq_set_wake request to the
ARM GIC interrupt controller, because there we've got a function
that saves and restores interrupts on suspend and resume.

This is strictly needed for correct functionality of the wakeup
interrupts.

Signed-off-by: Raphiel Rollerscaperers <rapherion@raphielgang.org>
Signed-off-by: Twisted Prime <36546624+TwistedPrime@users.noreply.github.com>